Restaurant Management Application project document
Developed a front-end management application for a restaurant that caters to both customers and management. The application allows customers to place food orders and choose dining options, while providing management with the ability to review and process these orders. The core functionality includes reading and writing files to handle order data using Java.